+++ tests/pyccel/test_pyccel.py
@@ -721,11 +721,6 @@ def test_print_integers(language):
     pyccel_test("scripts/print_integers.py", language=language, output_dtype=types)
 
 #------------------------------------------------------------------------------
-def test_print_tuples(language):
-    types = str
-    pyccel_test("scripts/print_tuples.py", language=language, output_dtype=types)
-
-#------------------------------------------------------------------------------
 def test_print_sp_and_end(language):
     types = str
     pyccel_test("scripts/print_sp_and_end.py", language=language, output_dtype=types)
@@ -1240,3 +1235,8 @@ def test_stubs(language):
         shutil.rmtree(wk_dir)
 
     assert expected_pyi == generated_pyi
+
+#------------------------------------------------------------------------------
+def test_builtin_container_print(language):
+    pyccel_test("scripts/print_builtin_containers.py", output_dtype = str,
+            language = language)