+++ CHANGELOG.md
@@ -86,6 +86,7 @@ All notable changes to this project will be documented in this file.
 -   \[INTERNALS\] Added `container_rank` property to `ast.datatypes.PyccelType` objects.
 -   \[INTERNALS\] Add a `__call__` method to `FunctionDef` to create `FunctionCall` instances.
 -   \[INTERNALS\] Allow the use of magic methods to describe container methods.
+-   \[INTERNALS\] Add a simplify method to `PyccelGt` for literals.
 -   \[DEVELOPER\] Added an improved traceback to the developer-mode errors for errors in function calls.
 -   \[DEVELOPER\] Added an environment variable to globally activate developer-mode for errors.
 -   \[DEVELOPER\] Added a spell checker for the code itself.
@@ -156,6 +157,7 @@ All notable changes to this project will be documented in this file.
 -   #2274 : Always use the C compiler to build the C wrapper for NumPy arrays (fixes Intel failures).
 -   #2285 : Reduce number of warnings in unit tests.
 -   #2295 : Fix wrapper handling of constant array arguments.
+-   #2097 : Fix printing of an empty list.
 
 ### Changed